+++ src/rockstor/smart_manager/models/share_replication.py
@@ -48,7 +48,7 @@ class ReplicaTrail(models.Model):
     """
     replica = models.ForeignKey(Replica)
     snap_name = models.CharField(max_length=1024)
-    kb_sent = models.IntegerField(default=0)
+    kb_sent = models.BigIntegerField(default=0)
     snapshot_created = models.DateTimeField(null=True)
     snapshot_failed = models.DateTimeField(null=True)
     send_pending = models.DateTimeField(null=True)
@@ -91,7 +91,7 @@ class ReceiveTrail(models.Model):
     """
     rshare = models.ForeignKey(ReplicaShare)
     snap_name = models.CharField(max_length=1024)
-    kb_received = models.IntegerField(default=0)
+    kb_received = models.BigIntegerField(default=0)
     receive_pending = models.DateTimeField(null=True)
     receive_succeeded = models.DateTimeField(null=True)
     receive_failed = models.DateTimeField(null=True)
